+++ src/uni_modules/wot-design-uni/components/wd-steps/wd-steps.vue
@@ -35,4 +35,3 @@ linkChildren({ props })
 <style lang="scss" scoped>
 @import './index.scss';
 </style>
-./type